MBS-10829: Fix SOLR syntax for recording with no length
This wasn't actually finding recordings without length at all, except
when the track also had no length (which made the query different).

This makes it so that recordings without length can also be found.
Because we're unlikely to want recordings without length and *also*
with a different artist as the track, and that can cause a lot of false
positives, this restricts the no-length results to ones
with matching artists.